Adding --input_shapes arg to control input
We were forcing static input shapes, as that is generally the rule of
thumb, but it turned out that allowing 2 or more input shapes may
actually be better in the end. It causes bigger overhead in the start,
but we're not wasting as many flops as a static shape;

If we always force 128 sentences and 64 tokens in one batch, then
sentences shorter than 64 tokens will be padded to 64.
However, if we allow 128x64 and 256x32 inputs, we both

* sometimes process more records in one step
* waste less flops because we have much less pads in the input.

In this pr, I:
* deleted example command in the docstring -- will add a fresh one later
* overrode `fairseq/data/data_utils.batch_by_size` w/ tpu version
* changed `collate_tokens_tpu`
* input shapes are controlled by --input_shapes
* changed `parse_args` accordingly. We now error when gpu input control
  flags are enabled.
* cleaned the log_step function